Add "Insert/Delete" as script context actions
The current way of tracking the current row and using it
as a script index is starting to show its limitations.
Ideally I'd do away with the script id thing, and just
rewrite the entire ScriptArchive class as a QAbstractListModel
with a doubly linked list of Script objects backing it
(instead of the static vector).